refactor: Card 객체 도메인으로 이동
